Implemented improvements from UAT + other fixes
Fixed:
- Analysis Images and graphs can be opened full screen
- Caching of images fixed, they will now change dynamically
- Navigation/get started prompt on home page added
- Twitter lookup no results crash/bug fixed + feedback for no results.
