+++ Flask_App/templates/analysis.html
@@ -125,10 +125,10 @@
 					<td width="25%">Word Frequency</td>
 				</tr>
 				<tr class="images">
-					<td class="images"><img src="/static/img_confMatrix.jpg" width="100%"></td>
-					<td class="images"><img src="/static/img_realCloud.jpg" width="100%"></td>
-					<td class="images"><img src="/static/img_fakeCloud.jpg" width="100%"></td>
-					<td class="images"><img src="/static/img_freqWords.jpg" width="100%"></td>
+					<td class="images"><a href="/static/img_confMatrix.jpg"><img src="/static/img_confMatrix.jpg" width="100%"></a></td>
+					<td class="images"><a href="/static/img_realCloud.jpg"><img src="/static/img_realCloud.jpg" width="100%"></a></td>
+					<td class="images"><a href="/static/img_fakeCloud.jpg"><img src="/static/img_fakeCloud.jpg" width="100%"></a></td>
+					<td class="images"><a href="/static/img_freqWords.jpg"><img src="/static/img_freqWords.jpg" width="100%"></a></td>
 				</tr>
 			</table>
 			<br>
