+++ Flask_App/templates/home.html
@@ -76,11 +76,18 @@
 		<br>
 		<br>
 		<div class='container'>
-		<p style=text-align:center>A fake news detection web application using
+		<p style=text-align:center><b>A fake news detection web application using
 			Machine Learning algorithms, deployed locally using the Flask Python
-			Framework. </p>
+			Framework. </b></p>
 			<br>
-		<p style=text-align:center style="padding:50px" style="text-align: justify;">With the widespread and ever-increasing use of and reliance on social media platforms, information, news and current affairs have become endlessly accessible on all platforms. Fake news itself, is the existence of false or misleading information presented to readers as legitimate news. Creators of fake news can have various motives, often intending to damage reputation of persons or entities, money making scams, political agendas or conspiracy theories. This misinformation can contain false and/or exaggerated claims, provoking readers to share the information, and potentially causing them to go viral due to the media propagation and sharing algorithms. Hence, it is important to recognise and ascertain the legitimacy of sources with the goal of protecting people and therefore wider society from those intending to mislead others. </p>
+		<p style=text-align:center style="padding:50px" style="text-align: justify;">With the widespread and ever-increasing use of and reliance on social media platforms, 
+			information, news and current affairs have become endlessly accessible on all platforms. Fake news itself, is the existence of false or misleading information 
+			presented to readers as legitimate news. Creators of fake news can have various motives, often intending to damage reputation of persons or entities, money making scams, 
+			political agendas or conspiracy theories. This misinformation can contain false and/or exaggerated claims, provoking readers to share the information, and potentially 
+			causing them to go viral due to the media propagation and sharing algorithms. Hence, it is important to recognise and ascertain the legitimacy of sources with the goal 
+			of protecting people and therefore wider society from those intending to mislead others. </p>
+		<p style=text-align:center style="padding:50px" style="text-align: justify;"> <br><i>Please select an input mode on the left to get started!
+			</p>
 		</div>
 	</div>
 </body>
